+++ src/FluxFlow.Components.Validation/Contracts/IJsonSchemaValueSelector.cs
@@ -1,8 +1,0 @@
-using FluxFlow.Components.Validation.Options;
-
-namespace FluxFlow.Components.Validation.Contracts;
-
-public interface IJsonSchemaValueSelector<TInput>
-{
-    object? Select(TInput input, JsonSchemaValidatorContext context);
-}